+++ start.js
@@ -17,6 +17,21 @@ if (!envExists) {
   .pipe(fs.createWriteStream(path.join(__dirname, '/.env')))
 }
 
+// Create template session data defaults file if it doesn't exist
+const dataDirectory = path.join(__dirname, '/app/data')
+const sessionDataDefaultsFile = path.join(dataDirectory, '/session-data-defaults.js')
+const sessionDataDefaultsFileExists = fs.existsSync(sessionDataDefaultsFile)
+
+if (!sessionDataDefaultsFileExists) {
+  console.log('Creating session data defaults file')
+  if (!fs.existsSync(dataDirectory)) {
+    fs.mkdirSync(dataDirectory)
+  }
+
+  fs.createReadStream(path.join(__dirname, '/lib/template.session-data-defaults.js'))
+  .pipe(fs.createWriteStream(sessionDataDefaultsFile))
+}
+
 // Run gulp
 const spawn = require('cross-spawn')